+++ x/finality/keeper/power_dist_change_test.go
@@ -96,7 +96,7 @@ func FuzzDistributionCache_BtcUndelegateSameBlockAsExpiration(f *testing.F) {
 
 		ctx = ProduceBlock(t, r, app, ctx)
 
-		for covI, _ := range covenantSKs {
+		for covI := range covenantSKs {
 			_, err = msgSrvrBtcStk.AddCovenantSigs(ctx, delCreationInfo.MsgAddCovenantSigs[covI])
 			require.NoError(t, err)
 
@@ -428,6 +428,7 @@ func FuzzProcessAllPowerDistUpdateEvents_Determinism(f *testing.F) {
 func CreateFpAndBtcDel(
 	t *testing.T,
 	r *rand.Rand,
+	addCovenantSigs bool,
 ) (
 	h *testutil.Helper,
 	del *btcstktypes.BTCDelegation,
@@ -448,7 +449,7 @@ func CreateFpAndBtcDel(
 
 	delSK, _, err := datagen.GenRandomBTCKeyPair(r)
 	h.NoError(err)
-	_, _, del, _, _, _, err = h.CreateDelegationWithBtcBlockHeight(
+	_, msg, del, _, _, _, err := h.CreateDelegationWithBtcBlockHeight(
 		r,
 		delSK,
 		fpPK,
@@ -462,6 +463,14 @@ func CreateFpAndBtcDel(
 		30,
 	)
 	h.NoError(err)
+
+	if addCovenantSigs {
+		// needed for covenant check in message handler
+		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 30})
+
+		h.CreateCovenantSigs(r, covenantSKs, msg, del, 30)
+	}
+
 	return h, del, covenantSKs
 }
 
@@ -470,7 +479,7 @@ func FuzzProcessAllPowerDistUpdateEvents_ActiveAndUnbondTogether(f *testing.F) {
 
 	f.Fuzz(func(t *testing.T, seed int64) {
 		r := rand.New(rand.NewSource(seed))
-		h, del, _ := CreateFpAndBtcDel(t, r)
+		h, del, _ := CreateFpAndBtcDel(t, r, true)
 
 		eventActive := btcstktypes.NewEventPowerDistUpdateWithBTCDel(&btcstktypes.EventBTCDelegationStateUpdate{
 			StakingTxHash: del.MustGetStakingTxHash().String(),
@@ -492,7 +501,7 @@ func FuzzProcessAllPowerDistUpdateEvents_ActiveAndSlashTogether(f *testing.F) {
 
 	f.Fuzz(func(t *testing.T, seed int64) {
 		r := rand.New(rand.NewSource(seed))
-		h, del, _ := CreateFpAndBtcDel(t, r)
+		h, del, _ := CreateFpAndBtcDel(t, r, true)
 
 		eventActive := btcstktypes.NewEventPowerDistUpdateWithBTCDel(&btcstktypes.EventBTCDelegationStateUpdate{
 			StakingTxHash: del.MustGetStakingTxHash().String(),
@@ -512,7 +521,7 @@ func FuzzProcessAllPowerDistUpdateEvents_PreApprovalWithSlahedFP(f *testing.F) {
 
 	f.Fuzz(func(t *testing.T, seed int64) {
 		r := rand.New(rand.NewSource(seed))
-		h, delNoPreApproval, covenantSKs := CreateFpAndBtcDel(t, r)
+		h, delNoPreApproval, covenantSKs := CreateFpAndBtcDel(t, r, false)
 
 		// activates one delegation to the finality provider without preapproval
 		eventActive := btcstktypes.NewEventPowerDistUpdateWithBTCDel(&btcstktypes.EventBTCDelegationStateUpdate{
@@ -597,7 +606,7 @@ func FuzzProcessAllPowerDistUpdateEvents_ActiveAndJailTogether(f *testing.F) {
 
 	f.Fuzz(func(t *testing.T, seed int64) {
 		r := rand.New(rand.NewSource(seed))
-		h, del, _ := CreateFpAndBtcDel(t, r)
+		h, del, _ := CreateFpAndBtcDel(t, r, false)
 
 		eventActive := btcstktypes.NewEventPowerDistUpdateWithBTCDel(&btcstktypes.EventBTCDelegationStateUpdate{
 			StakingTxHash: del.MustGetStakingTxHash().String(),
@@ -622,7 +631,7 @@ func FuzzProcessAllPowerDistUpdateEvents_SlashActiveFp(f *testing.F) {
 	f.Fuzz(func(t *testing.T, seed int64) {
 		t.Parallel()
 		r := rand.New(rand.NewSource(seed))
-		h, del, _ := CreateFpAndBtcDel(t, r)
+		h, del, _ := CreateFpAndBtcDel(t, r, false)
 
 		eventActive := btcstktypes.NewEventPowerDistUpdateWithBTCDel(&btcstktypes.EventBTCDelegationStateUpdate{
 			StakingTxHash: del.MustGetStakingTxHash().String(),
@@ -1784,4 +1793,48 @@ func TestIgnoreExpiredEventIfThereIsNoQuorum(t *testing.T) {
 	// as delegation does not have covenant quorum
 	h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(btcTip).AnyTimes()
 	h.BeginBlocker()
+}
+
+func TestIgnoreUnbondingEventIfThereIsNoQuorum(t *testing.T) {
+	t.Parallel()
+	r := rand.New(rand.NewSource(time.Now().UnixNano()))
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+
+	// mock BTC light client and BTC checkpoint modules
+	btclcKeeper := btcstktypes.NewMockBTCLightClientKeeper(ctrl)
+	btccKeeper := btcstktypes.NewMockBtcCheckpointKeeper(ctrl)
+	h := testutil.NewHelperNoMocksCalls(t, btclcKeeper, btccKeeper)
+
+	// set all parameters
+	h.GenAndApplyParams(r)
+
+	_, fpPK, _ := h.CreateFinalityProvider(r)
+
+	delSK, _, err := datagen.GenRandomBTCKeyPair(r)
+	h.NoError(err)
+	_, _, del, _, _, _, err := h.CreateDelegationWithBtcBlockHeight(
+		r,
+		delSK,
+		fpPK,
+		int64(2*10e8),
+		1000,
+		0,
+		0,
+		false,
+		false,
+		10,
+		30,
+	)
+	h.NoError(err)
+	eventUnbond := btcstktypes.NewEventPowerDistUpdateWithBTCDel(&btcstktypes.EventBTCDelegationStateUpdate{
+		StakingTxHash: del.MustGetStakingTxHash().String(),
+		NewState:      btcstktypes.BTCDelegationStatus_UNBONDED,
+	})
+	events := []*btcstktypes.EventPowerDistUpdate{eventUnbond}
+
+	// k.IncentiveKeeper.BtcDelegationUnbonded(ctx, fp, del, sats) won't be called
+	// as delegation does not have covenant quorum
+	newDc := h.FinalityKeeper.ProcessAllPowerDistUpdateEvents(h.Ctx, ftypes.NewVotingPowerDistCache(), events)
+	require.Len(t, newDc.FinalityProviders, 0)
 }